feat: guest 들 간의 실시간 poll 상태 공유 기능 추가
reducer 함수도 별도의 파일로 분리함
